Add validation for new domains
